+++ prisma/schema.prisma
@@ -48,12 +48,13 @@ model Session {
 model User {
   id            String    @id @default(auto()) @map("_id") @db.ObjectId
   name          String?
-  firstName     String?
-  lastName      String?
-  pronouns      String?
+  firstName     String    @default("")
+  lastName      String    @default("")
+  pronouns      String    @default("")
   email         String?   @unique
   emailVerified DateTime?
   image         String?
+  position      String?
 
   // @deprecated Pending for removal when we expand to multiple universities
   admin Boolean @default(false)
