+++ src/app/api/user/[uid]/edit-profile/route.schema.ts
@@ -1,16 +1,17 @@
 import { z } from "zod";
-import { Prisma } from "@prisma/client";
 
 /*
   firstName     String
   lastName      String
   pronouns      String
+  position      String
 */
 
 export const EditProfileRequest = z.object({
   firstName: z.string(),
   lastName: z.string(),
   pronouns: z.string(),
+  position: z.string(),
 });
 
 export const EditProfileResponse = z.discriminatedUnion("code", [
@@ -21,25 +22,6 @@ export const EditProfileResponse = z.discriminatedUnion("code", [
   z.object({
     code: z.literal("INVALID_FORM"),
     message: z.literal("Invalid form submission"),
-  }),
-  z.object({
-    code: z.literal("UNKNOWN"),
-    message: z.literal("Unknown error received"),
-  }),
-]);
-
-export const GetProfileRequest = z.object({
-  id: z.string(),
-});
-
-export const GetProfileResponse = z.discriminatedUnion("code", [
-  z.object({
-    code: z.literal("SUCCESS"),
-    message: z.literal("Profile successfully retrieved"),
-  }),
-  z.object({
-    code: z.literal("INVALID_ID"),
-    message: z.literal("Invalid id"),
   }),
   z.object({
     code: z.literal("UNKNOWN"),
